feat: replace and rename `organization` artefacts with `builder`
